Tracks (#26)
* [WIP] Tracks animations

* New utility function added

get_angle function to return the angle from vector
get_dummy_texture function moved to TestUtility.hpp, now it returns
std::unique_ptr<sf::Texture> to make it reusable in other test files

* First attempt to implement TracesHandler

* TracesHandler can handle adding multiple moves with different angle

* TracesHandler can handle position setting for single trace

Trace origin is set to middle top point. After each move the trace is
positioned using tank sprite middle bottom point transform.

TODO:
- in case we move in single direction and trace exceeds maximum sprite
height, we must also find position for the trace continuation

* Tank traces functional when moving forward

When tank is moving forward the traces are displayed well (no breaks, sprite count is optimized by extending each sprite height)
When tank is changing direction and still moving forward new sprites are added with correct size and angle

TODO:
- fix tests
- fix traces display when tank is moving backwards
- add some kind of a counter to remove old traces

* Move angle calcuation fix - TracesHandler is functional

TracesHandler:
- correctly adds traces when the tank is moving forward,
- correctly adds traces when the tank is changing the move angle (turns left, turns right).
- correctly adds traces when the tank is moving backward (the turns also work correctly).
- correctly extends sprite height in order to minimize draw function calls

TODO:
- TracesHandler does not remove old traces
- TracesHandler should be parametrized (for example the parameter which indicates how long the tracks will be visible)
- TracesHandlerTests must be rewritten
- TracesHandler code must be refactored and cleaned up
- naming convention should be verified for consistency
- code should be formatted before push

* TracesHandlerTest rework & files formatting

Tests reworked; single test verifies single functionality
added new method to TracesHandler to get the maximum texture height allowed by handler

* New folders added to .gitignore

* TracesHandler refactored

- get_sprite function encapsulates trace creation
- get_opposite_angle handles angle calculation for traces
- sprite position is set with offset equal to available sprite height instead of entire move distance value
- functions used to calculate new texture height in case our current move sprite height + last trace height exceeds max texture height
- simplified update function by reorganizing code
- removed unused code
- moved default Tank::~Tank() implementation to .cpp file so it is possible to use forward declaration for TracesHandler in Tank.hpp

* Trace class added

Trace can be used instead of sf::Sprite. It is more lightweight, because Trace does not need to be transformed once it is positioned, the only things that are necessary is: increase / decrease the trace length. Thanks to using sf::VertexArray it is possible to easily decrease the trace length, which can be used to decay the traces

* TracesHandler reimplemented

TracesHandler now uses new Trace implementation, thanks to which 50% of code could be removed.

* TracesHandler parametrized

New parameters added to control TracesHandler behaviour:
- max trace age (how long before tracks start to dissapear)
- trace decay rate (how fast the tracks are dissapearing)

* TracesHandler decay tests added

New test cases:
- veryfing that trace height is decreasing
- veryting that trace will be removed when decay value is greater than texture height

TracesHandler parameters moved to struct, so it is more readable to create TracesHandler using designated initializers.

Minor improvement in header management (SquareRootEngine.hpp include removed from Board.hpp)

* Formatting fix

* Minor CI code fixes

- uint typedef removed, because it does not exist on windows
- SquareRootEngine max_speed parameter type changed from int to float

* Further CI windows code fixes

* Removing uint typedef because windows doesnt support it

* Remove unnecessary headers

Fixed compilation warning when int was implicitly casted to float

+++ src/SquareRootEngine.hpp
@@ -4,7 +4,7 @@
 
 class SquareRootEngine : public Engine {
  public:
-  SquareRootEngine(const int step_count, const int max_speed);
+  SquareRootEngine(const int step_count, const float max_speed);
   void set_gear(const Gear gear) override;
   float get_current_speed() const override;
   sf::Vector2f get_position_delta(const float rotation_radians) override;
